Modularize map item cleanup
