+++ game_core/playscreen_components/player_system/player_character.py
@@ -4,7 +4,7 @@ Player Character module - implements the playable character with animations
 import pygame
 import math
 from settings import *
-from .animation_handler import blit_aligned
+from .player_animation import blit_aligned
 from .sprite_loader import load_character_sprites
 
 class PlayerCharacter(pygame.sprite.Sprite):
